LayoutTests/streams/readablestream-pull-source.html

<!DOCTYPE html>
<script src="resources/streams-utils.js"></script>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>

t3 = async_test('ReadableStream is able to pull data repeatedly if it\'s available synchronously');
t3.step(function() {
    var i = 0;
    var rs = new ReadableStream({
        pull: function(enqueue, close) {
            if (++i <= 10) {
                enqueue(i);
            } else {
                close();
            }
        }
    });
    rs.ready.then(t3.step_func(function() {
       var data = [];
        while (rs.state === 'readable') {
            data.push(rs.read());
        }
        assert_array_equals(data, [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]);
        t3.done();
    }));
});

</script>
